requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
numpy<2
requests==2.31.0

# Machine Learning
faiss-cpu==1.8.0
torch>=1.13.0
scikit-learn==1.3.0

# NLP and Sentiment
textblob==0.17.1
vaderSentiment==3.3.2

# Graph Analysis
networkx==3.1

# Development and Testing
pytest==7.4.2
pytest-mock==3.14.1
httpx==0.25.0

# Note: Ollama CLI must be installed separately
# Windows: winget install Ollama.Ollama
# macOS: brew install ollama
# Linux: curl -fsSL https://ollama.ai/install.sh | sh
google-generativeai==0.3.1
beautifulsoup4==4.12.3
